Updates to try the particle method
