+++ method_comparison_results/method_comparison_results.csv
@@ -1,4 +1,4 @@
-,original,multivariate
-beta error:,0.014095807662269033,0.04845801673502992
-alpha error:,7.720105912629026,5.5197868919506465
-time_elapsed,236.58,77.21
+,original,multivariate,approximation
+beta error,0.009840430228705621,0.032144645361144475,0.0025760789164722686
+alpha error,8.24206155095091,4.992357546928772,10.590235997262043
+time_elapsed,270.6,64.87,497.83
